+++ bootstrap/freebsd.sh
@@ -1,6 +1,6 @@
 #!/bin/sh -xe
 
-pkg install -Ay \
+"$SUDO" pkg install -Ay \
   git \
   python \
   py27-virtualenv \
